change type back

+++ embedded8266/user/adc.c
@@ -51,7 +51,7 @@ uint16 hs_adc_read(void)
 {
    uint16 sardata[8];
    uint8 i;
-	uint32_t sar_dout = 0;
+	uint16_t sar_dout = 0;
 
     while (GET_PERI_REG_BITS(0x60000D50, 26, 24) > 0); //wait r_state == 0
     read_sar_dout(sardata);
@@ -76,7 +76,7 @@ uint16 hs_adc_read(void)
     SET_PERI_REG_MASK(0x60000D50, 0x02);    //force_en=1
 #endif
 
-    return (uint16)sar_dout;      //tout is 10 bits fraction
+    return sar_dout;      //tout is 10 bits fraction
 }
